Convert tools/maven:pom_file to be a native skylark rule instead of a macro
The macro was awkward and shelled out to a few things that can be replaced with native bazel actions. This also makes it more platform-independent instead of relying on the default python version. It also avoids writing a .deps_xml file that was only used to be passed as an impelmentation detail to another action.

The one "downside" to this rule is that it doesn't indent the dependencies XML block, but that doesn't seem so bad. Users can always call xmllint on this file themselves if they choose.

Closes https://github.com/google/bazel-common/pull/9

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=193038233