Update bazel_common to org.pantsbuild:jarjar:1.7.2
This fixes bazel_common's `jarjar_library` which I think was broken when updating to `org.ow2.asm:asm-commons:7.2`.

The issue is that `org.pantsbuild:jarjar:1.6.3` was depending on a class that is removed in the new version of asm-commons:

  java.lang.NoClassDefFoundError: org/objectweb/asm/commons/RemappingClassAdapter

so we need to also update pantsbuild.

I've verified locally that upgrading from 1.6.3 to 1.7.2 fixes the `jarjar_library` for Dagger.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=290281067