Adjust usage of 'find' based on Mac or Linux host
The find command in Mac is the BSD version while the one on Linux is GNU and they
differ in capabilities, including the ability to find files with regex. On Mac and extra argument has to be passed, '-E' to interpret regular expressions with '-regex'. Meanwhile BSD regex interpretation also slightly differs from GNU, specifically it seems in the GNU version the parenthesis for a match group have to be escaped.

This change updates the jarjar command to take the above into account, using 'uname' to check if host is Darwin.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=310938373