Construct the `-sourcepath` argument of `javadoc_library` using the srcs
The existing approach just looks for the first directory name `.*java` in the
working directory which could very well be wrong.

Instead we'll assume the root packages declared will appear as subpaths in the
srcs. We'll then take the unique set of parent directories of these subpaths and
use that as the `-sourcepath`.

PiperOrigin-RevId: 666325391