Bump `failureaccess` to 1.0.3.
Bazel is already automatically pulling 1.0.3, since that's a transitive dependency of recent Guava releases. This just brings our explicit specification in line with that.

(It's possible that it would make sense for us to stop listing the `failureaccess` dependency entirely: We're always going to be getting it as part of Guava. But maybe that would prevent us from exporting it from the `guava` target that we define? I don't feel like figuring that out right at this moment.)

(followup to [our modularization efforts](https://github.com/google/guava/issues/2970))

RELNOTES=n/a
PiperOrigin-RevId: 741553352